Consolidate commerce execution in Eve tools
