+++ apps/template/agent/tools/add-cart-note.ts
@@ -1,10 +1,11 @@
 import { defineTool } from "eve/tools";
+import { z } from "zod";
 
-import { commerceSchemas } from "../../lib/agent/commerce";
-import { callCommerce } from "../lib/commerce";
+import { getSessionCartId, mutateCart } from "../lib/cart";
+import { runCommerce } from "../lib/commerce";
 
 export default defineTool({
   description: "Attach a note to the shopper's current cart only when asked.",
-  inputSchema: commerceSchemas["add-cart-note"],
-  execute: (input, ctx) => callCommerce("add-cart-note", input, ctx),
+  inputSchema: z.strictObject({ note: z.string() }),
+  execute: ({ note }, ctx) => runCommerce(() => mutateCart(getSessionCartId(ctx), { note })),
 });
