+++ apps/template/agent/tools/add-to-cart.ts
@@ -1,11 +1,18 @@
 import { defineTool } from "eve/tools";
+import { z } from "zod";
 
-import { commerceSchemas } from "../../lib/agent/commerce";
-import { callCommerce } from "../lib/commerce";
+import { getSessionCartId, mutateCart } from "../lib/cart";
+import { runCommerce } from "../lib/commerce";
 
 export default defineTool({
   description:
     "Add a confirmed ProductVariant ID to the shopper's cart. Only when asked; never retry an uncertain change.",
-  inputSchema: commerceSchemas["add-to-cart"],
-  execute: (input, ctx) => callCommerce("add-to-cart", input, ctx),
+  inputSchema: z.strictObject({
+    quantity: z.number().int().min(1).max(99).default(1),
+    variantId: z.string().regex(/^gid:\/\/shopify\/ProductVariant\/[0-9]+$/),
+  }),
+  execute: ({ quantity, variantId }, ctx) =>
+    runCommerce(() =>
+      mutateCart(getSessionCartId(ctx), { lines: [{ merchandiseId: variantId, quantity }] }),
+    ),
 });
