+++ apps/template/agent/tools/browse-collection.ts
@@ -1,10 +1,22 @@
 import { defineTool } from "eve/tools";
+import { z } from "zod";
 
-import { commerceSchemas } from "../../lib/agent/commerce";
-import { callCommerce } from "../lib/commerce";
+import { toAgentProduct } from "../../lib/agent/products";
+import { fetchCollectionProducts } from "../../lib/shopify/catalog/server";
+import { productHandleSchema } from "../lib/catalog";
+import { runCommerce } from "../lib/commerce";
 
 export default defineTool({
   description: "Browse products by collection handle from list-collections or page context.",
-  inputSchema: commerceSchemas["browse-collection"],
-  execute: (input, ctx) => callCommerce("browse-collection", input, ctx),
+  inputSchema: z.strictObject({
+    collection: productHandleSchema,
+    sortKey: z
+      .enum(["best-matches", "price-low-to-high", "price-high-to-low", "BEST_SELLING", "CREATED"])
+      .default("best-matches"),
+  }),
+  execute: ({ collection, sortKey }) =>
+    runCommerce(async () => {
+      const { products } = await fetchCollectionProducts({ collection, limit: 12, sortKey });
+      return { products: products.map(toAgentProduct) };
+    }),
 });
