+++ apps/template/agent/tools/get-product-details.ts
@@ -1,11 +1,20 @@
 import { defineTool } from "eve/tools";
+import { z } from "zod";
 
-import { commerceSchemas } from "../../lib/agent/commerce";
-import { callCommerce } from "../lib/commerce";
+import { toAgentProductDetails } from "../../lib/agent/products";
+import { fetchProductWithVariants } from "../../lib/shopify/catalog/server";
+import { productHandleSchema } from "../lib/catalog";
+import { runCommerce } from "../lib/commerce";
 
 export default defineTool({
   description:
     "Get details for a specific product and render its interactive variant picker. Use for product questions or purchase selection, not to display initial search-result cards.",
-  inputSchema: commerceSchemas["get-product-details"],
-  execute: (input, ctx) => callCommerce("get-product-details", input, ctx),
+  inputSchema: z.strictObject({ handle: productHandleSchema }),
+  execute: ({ handle }) =>
+    runCommerce(async () => {
+      const product = await fetchProductWithVariants({ handle });
+      return product
+        ? { product: toAgentProductDetails(product) }
+        : { error: "Product not found." };
+    }),
 });
