+++ apps/template/agent/tools/get-recommendations.ts
@@ -1,10 +1,28 @@
 import { defineTool } from "eve/tools";
+import { z } from "zod";
 
-import { commerceSchemas } from "../../lib/agent/commerce";
-import { callCommerce } from "../lib/commerce";
+import { toAgentProduct } from "../../lib/agent/products";
+import { fetchComplementaryProducts, fetchRelatedProducts } from "../../lib/shopify/catalog/server";
+import { productHandleSchema } from "../lib/catalog";
+import { runCommerce } from "../lib/commerce";
 
 export default defineTool({
   description: "Show related and complementary products for a product handle.",
-  inputSchema: commerceSchemas["get-recommendations"],
-  execute: (input, ctx) => callCommerce("get-recommendations", input, ctx),
+  inputSchema: z.strictObject({ handle: productHandleSchema }),
+  execute: ({ handle }) =>
+    runCommerce(async () => {
+      const [related, complementary] = await Promise.all([
+        fetchRelatedProducts({ handle }),
+        fetchComplementaryProducts({ handle }),
+      ]);
+      return {
+        products: [
+          ...new Map(
+            [...related, ...complementary].map((product) => [product.id, product]),
+          ).values(),
+        ]
+          .slice(0, 12)
+          .map(toAgentProduct),
+      };
+    }),
 });
