+++ apps/template/agent/tools/list-collections.ts
@@ -1,10 +1,18 @@
 import { defineTool } from "eve/tools";
+import { z } from "zod";
 
-import { commerceSchemas } from "../../lib/agent/commerce";
-import { callCommerce } from "../lib/commerce";
+import { fetchCollections } from "../../lib/shopify/catalog/server";
+import { runCommerce } from "../lib/commerce";
 
 export default defineTool({
   description: "List this store's collection handles, titles and descriptions.",
-  inputSchema: commerceSchemas["list-collections"],
-  execute: (input, ctx) => callCommerce("list-collections", input, ctx),
+  inputSchema: z.strictObject({}),
+  execute: () =>
+    runCommerce(async () => ({
+      collections: (await fetchCollections()).map(({ description, handle, title }) => ({
+        description,
+        handle,
+        title,
+      })),
+    })),
 });
