+++ apps/template/agent/tools/present-products.ts
@@ -1,11 +1,22 @@
 import { defineTool } from "eve/tools";
+import { z } from "zod";
 
-import { commerceSchemas } from "../../lib/agent/commerce";
-import { callCommerce } from "../lib/commerce";
+import { fetchProductsByIds } from "../../lib/shopify/catalog/server";
+import { matchingProducts, productOptionsSchema } from "../lib/catalog";
+import { runCommerce } from "../lib/commerce";
 
 export default defineTool({
   description:
     "Render chosen exact Shopify Product IDs from native catalog search. Filter only by product options explicitly requested by the shopper; otherwise use []. Never infer preferences from featured variants. Other authored product tools already render their results.",
-  inputSchema: commerceSchemas["present-products"],
-  execute: (input, ctx) => callCommerce("present-products", input, ctx),
+  inputSchema: z.strictObject({
+    ids: z.array(z.string().regex(/^gid:\/\/shopify\/Product\/[0-9]+$/)).max(12),
+    options: productOptionsSchema,
+  }),
+  execute: ({ ids, options }) =>
+    runCommerce(async () => ({
+      products: await matchingProducts(
+        ids.length ? await fetchProductsByIds({ ids }) : [],
+        options,
+      ),
+    })),
 });
